Minima 1.1 pre-release 2
A simple(r) way to install themes - upload the stylesheet to /and/themes
and change the link in the database via the admin interface.
Addons now work - upload it to ant/addons and add it (with <?php include
'ant/addons/someaddon.php'; ?>)
something else, boring stuff aka I forgot to put it into the main
changelog.